backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
httpx>=0.27.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.40.0
chronos-forecasting>=2.2.0
google-genai>=1.0.0
